Refs[TASK][40972] Refactor code
